use two different images for large vs small viewports
Also, this commit removes some javascript that is no longer necessary. The deleted code would have hard-coded the layout container size when the viewport was larger than the image. That is no longer necessary since desktop viewports will receive a large enough image to cover the full screen for 99% of screens (according to some data on w3schools).