clean up music page
Remove the background image on the music page. Center text content on the music page in a container that takes up the full screen width on small devices and 70% of the screen width on large devices. Remove javascript for image container sizing that is no longer used on the music page.