Remove navbar gray line display on small screens
When the viewport was small (for example, 360 px on a Galaxy S5), the gray lines next to the navbar would still appear but would be small and a bit awkward-looking. This commit applies changes so that if the viewport width is less than or equal to 360 px, the gray lines will not be displayed.