Speed up CSS fade-in transition
Also, remove some unneeded CSS. I have decided to stick with the current animation implementation, which immediately unmounts a Route component when another route is clicked, and immediately mounts and starts to fade in the Route component that was clicked. I looked in to a different implementation which would have simultaneously faded out one component and faded in its replacement, but I gained an impression that the two components would kind of visually interfere with each other and compete with each other. So, I am going to just stick with what I have now instead of implementing a simultaneous fade-out / fade-in look.