Rename a component for better clarity
Rename the PageContent component to Routes. This is more clearn and explicit, since the purpose of the component is to encapsulate the display of the selected Route component. There is already another component called Content which wraps the actual content itself, so naming the PageContent component as such was redundant anyway.