Extract setting selected link to a dedicated method
Previously, the state indicating an active link was set in the JSX, which made it a bit difficult to find the logic for setting state. This commit implements a method setActive(link_name) which sets the active link state when a link is clicked. Keeping this logic in a dedicated method makes it more directly evident where the state is being set.