Implement material design with custom CSS
This commit removes the library Material-UI and instead implements material design with custom CSS. This styling visually indicates component state and user interaction by using shadows to suggest height.

A navbar link that is not active has a small blurred shadow, which indicates low height. When the cursor hovers over a navbar link, the shadow is slightly bigger and more blurred, which visually indicates slightly higher height. When the cursor clicks on a link, the shadow is at its biggest and most blurred, which visually indicates high height and suggests that the link is active.