Improve visual distinction of interactions with links
This commit implements better visual feedback to the user with regards to their interaction with navbar links. A link can be inactive and not hovered over, inactive and hovered over, active and not hovered over, or active and hovered over. Each of these interaction states corresponds to a certain height that is suggested with shadowing. This improves the user experience by always providing visual feedback to the user that a link is being hovered over or not, while also always providing visual feedback to the user that a link is active or not.