Improve perception of speed on the page
This commit makes the page's response to user interaction seem faster. The only change in this commit is to speed up the animated transition between different navbar link states.

I interacted with the Google homepage and realized that the animated transition for a mouseover event over the search bar is very quick. So, I sped up the transition between this page's navbar link states to closer match Google's implementation.