Highlight the active link when a user types the URL
This commit fixes a bug where the visually highlighted active link would not change when the user manually typed a URL into the browser bar.

With these changes, when the user enters a URL for a certain location in their browser bar, the visually highlighted active link will update to reflect that location.