Implement active link highlighting for browser back/forward
This commit simplifies the codebase by entirely removing state (which previously stored the particular link that was active, and did not know about browser back/forward happening). Now, the active link is determined at the moment the navbar component is rendered by querying the browserHistory object. The navbar component finds out what the current route is, and visually highlights the navbar link representing that route.

This fixes a bug where the browser back/forward buttons were not notifying the application state of the change to the current route. By removing state and determining the active highlighted link immediately when the navbar is rendered, the browser can maintain its responsibility of remembering the browser history and the application can update its active link highlighting at any route the browser decides to visit.