Remove a redundant alt attribute in an image tag
There was an image tag with an alt attribute containing the word 'Picture', which resulted in the following warning: Redundant alt attribute. Screen-readers already announce `img` tags as an image. You don't need to use the words `image`, `photo,` or `picture` (or any specified custom words) in the alt prop

This commit removes the redundant word in the alt attribute, addressing the warning.