Improve visual display on Projects page
I received feedback that the images on the projects page were too large, and it was hard to tell where one project ended and another one started. This commit makes the images smaller, and uses alternating alignment of text and images for different projects in order to differentiate them from each other.

Also, this commit makes a few small changes to the copy on the Projects page.