Update browser history for React-Router v4
React-Router v4 does not support a browserHistory object, which the browser back and forward functionality of this app previously relied on. This commit applies changes that use the createBrowserHistory() method of the npm package 'history' for that purpose.

To enable these changes, this commit installs the npm packages 'history' and its dependency 'ajv'.