get data for the projects page from a backend API
this commit sets up the top-level application component to request data from a local server once the component has mounted. when the application receives the response, it stores the data in state and passes it down the component tree as props. the data is passed down through the Router component, and then through the Content component which renders the appropriate page component.

at this point, the only data being processed in this way is for the Projects page, and all of the other content is still hard-coded.

next steps might include refactoring the logic that handles how the data is displayed, replacing the local server endpoint with a production endpoint, and getting all of the data for other pages from the API as well.