Fix hyperlink functionality
This commit fixes a bug that was introduced when the text started coming from the back-end API instead of being hard-coded in the front-end source code. The bug was that the browser would interpret hyperlinks as plaintext insted of HTML, so the hyperlinks wouldn't redirect the browser to anywhere when clicked.

An update has been made to the API that distinguishes URLs from sentences within a paragraph. This commit updates the front-end application to consume the new API format and generate a clickable hyperlink from a URL that arrives via the API.