Fix link wrapping in small browser viewports
There was a bug where long HTML links weren't wrapping when the user shrunk the browser viewport. This caused the entire parent HTML element containing the link to not shrink when the viewport got smaller than the link width, resulting in content being cut off from the sides of the UI.

This commit fixes that bug by forcing links to wrap when the parent HTML element is shrunk. Now, the link and its parent HTML element re-size appropriately when the user re-sizes the browser viewport.