Integrate Pages API into Projects page
The back-end API has been updated so that the top-level object in the JSON response is a collection of pages, which can contain other pages in addition to the Projects page. This commit sets up the Projects page to integrate this new API format, without yet changing the client-side application for any of the other pages.