Refactor duplicated functionality
This commit abstracts some functionality that was duplicated in the Projects and About components. This should make the code easier to maintain and change.

Previously, there was a set of functions for processing the API data and generating HTML templates that existed with slightly different wording in both the Projects and About components. This functionality now exists in a new DataFormatter component that is imported into the Projects and About components. Also, there were duplicated CSS styles that are now found only in Layout.css.