Add an alt prop to image elements
Previously, image elements in the HTML did not have an alt prop of text to display if the image didn't load, resulting in this warning in the console: "img elements must have an alt prop, either with meaningful text, or an empty string for decorative images"

This commit adds an alt prop that is just an empty string, just because an image's only attribute as of now is a URL that wouldn't be meaningful to display to a user. This gets rid of the warning message but doesn't result in any meaningful text being displayed to a user if an image doesn't load in the browser.